Implement a simple run-time tracing mechanism
Building with the ERIS_TRACE=1 preprocessor symbol defined enables
building the tracing system. Even when built, it is disabled by
default, and the ERIS_TRACE environment variable must be defined
(and not empty) to enable the trace. Overhead is minimal, and no
function calls to eris_trace() are done if tracing is built but
disabled. If tracing support is not compiled in, no additional
code is ever generated by the compiler.