harness: Implement Unicode terminal output formatter
Add a new Utf8TermOutput formatter, which uses Unicode characters
and ANSI CSI color escape sequences to provide a neater output.

Also, this patch adds the logic needed to choose the "uterm" output
formatter when output is sent to a supported terminal, or "tap"
otherwise (e.g. when piping output to a file or other program, or
to dumber terminals).