Resolve typedef DIEs to their actual type
The DIE for a DW_AT_type can have a DW_TAG_typedef tag, in which case it owns
a nested reference to another DW_AT_type DIE, which is the actual type to use.
Note that this can happen for typedefs of typedefs, so resolving the actual
base type must be done recursively.

Also, the code to obtain the value of a DIE attribute which happens to be a
reference to another DIE is factored out in die_get_die_reference_attribute().