Remove code used to find library base address
The base address of loaded libraries is not needed because dlsym() is
being used to obtain the effective load address for a symbol. This
incurs in doing a double lookup, but works and it's simple enough.
If, later on, it would be desirable to pick the offset of a variable
or entry point from the DWARF DIEs, this commit can be reverted.