Do not store setter/getter pointers in ErisVariable
Provided that the setter and getter are never going to change for a given
type, it is possible to just use a pointer to the structure containining
the mapping of type infos to getter/setter functions. This will also allow
some extra niceties in the future like e.g. reflection for type names of
variables.