Support for runtime checks selectable at compile-time
This adds a number of CHECK*() macros, which generate additional run-time
checks. The checks can be compiled-in or left out of the build, by passing
the "--enable-checks" or "--disable-checks" parameters to the "configure"
script. Execution is aborted unconditionally when a runtime check fails, so
it is only recommended to enable them during development.