Convert ErisTypeInfo into an opaque struct
This patch moves the actual definition of ErisTypeInfo is moved into a new
eris-typing.c file (which effectively makes it opaque), and adds accessor
functions for the struct fields, with public prototypes in eris-typing.h.

Having a more self-contained ErisTypeInfo makes it easier to make sure that
instances are created from a single place, and that they are treated as
immutable by the rest of the code. This is a convenient change to implement
the caching of ErisTypeInfo values later on.