Support libdwarf headers in system include directory
Detect whether the libdwarf headers are installed in the system include
directory, and can be accessed as <libdwarf.h>, or whether they are present
in a subdirectory and accessible as <libdwarf/libdwarf.h>. Some distributions
(e.g. Debian and derivatives) do not have a separate directory for the
libdwarf headers, and some other do (e.g. Arch Linux).