Support building with bundled libdwarf and/or Lua
The build system will manage to download and build libdwarf and/or Lua,
and making a minimal build of them, if requested. To enable this feature,
use "configure --enable-bundled-lindwarf" or "configure --enable-bundled-lua"
(passing both arguments is supported as well).

Using bundled versions may be desirable in case the system-provided Lua
or libdwarf libraries cannot be used. For example, Debian and Ubuntu only
provide static versions of libdwarf which are compiled without -fPIC, so
they cannot be linked into Lua modules (which are dynamic libraries
themselves, and therefore PIC objects).

This adds in "wget", "tar", and "gzip" as additional tools required for
building.