Implement indexing of ErisVariable userdatas
This patch defines metamethods __index and __newindex, which allow to
access values of ErisVariable userdatas, including the ones wrapping
arrays. In the later case, indexes other than 1 can be used. The existing
methods to retrieve the variable name, its library, and type are changed
into attributes prefixed with double underscore:

- variable:name() becomes variable.__name
- variable:type() becomes variable.__type
- variable:library() becomes variable.__library

Additionally, the value can be accessed using "variable.__value",
which is effectively equivalent to "variable[1]". The attributes
prefixed with double underscore are read-only, so the __newindex
metamethod does not handle them.