Implement per-library global types cache
Add a hashtable-based type cache for each ErisLibrary. Type names in C
are actually completely global, but the DWARF debug information stores
relative offsets for each loaded shared object, and the offsets are used
as hashtable keys. The reason is that all nested type informations (e.g.
for structs) or referenced from function DIEs are offset in the global
types table, so using the offset makes the cache useable for all typeinfo
lookups. At least inside the same shared object.