Use cache to look up types of ErisVariable userdatas
This implements a revamped representation of ErisTypeInfo as well,
which uses nested synthetic types to encode some information (e.g.
const-ness, arrays, pointers) in a fashion similar to how DWARF does,
with tagged type infos which contain a reference to their base type
infos.

This means that more ErisTypeInfo references can point to elements
cached in the per-library type cache, and that the way of building
ErisTypeInfo values is more "constructive" (in a functional sense).