Support building typeinfos for opaque structs
Opaque structs are those which are typedef'd, but their members are
not actually known, e.g. in a header:

  typedef struct _Opaque Opaque;
  extern Opaque* opaque_new (void);

This type of structs have a DW_AT_byte_size of zero, and no DW_TAG_member
child DIEs. They can have an optional DW_AT_declaration flag, which is not
checked currently.

This patch allows for creation of typeinfos for this kind of structs, and
adds a eris_typeinfo_struct_is_opaque() query function.