Support union typeinfos, more robust member extraction
Implements support for ErisTypeInfo values that describe "union" types.
As the fields needed in members for them are the same as for structs,
structs are renamed as "compound" types, and unions are considered to
be "compound" as well.

Also, base types are now defined as data constants in eris-typing.c, to
avoid having to allocate memory from the heap when using them.