Introduce DW_TRACE_DIE() and DW_TRACE_DIE_ERROR()
Those macros print tracing messages, but they also print a DWARF DIE
in the form "<DIE offset name>", with their offset and name (or, if
not available, its tag name) alongside with the line sent to the
trace log. The macro uses dw_die_repr(), and it takes care of freeing
resources itself.