Do not use eris_library_get_die_ref_attribute_offset() directly
The function eris_library_get_die_ref_attribute_offset() is intended to obtain
the offset of a DIE in order to perform a lookup in the type cache. Provided
that the eris_library_fetch_die_type_ref_cached() can obtain the offset and
do the cache lookup itself, using it results in shorter code that is easier
to understand.