Handle pointers in cvalue_push()
This adds handling of pointers to the cvalue_push() function. The way
pointers are handled is analogous to C: a pointer is treated as an
array, but contrary to arrays, items accessed through a pointer are
not bounds-checked. This may change in the future so only getting the
pointed-to item can be accessed (i.e. a pointer dereference). Also,
NULL is mapped to "nil".