Support getting typeinfos for opaque pointers
TUEs which represent an opaque pointer type (i.e. that points to an
unknown type, but they are known to be pointers so they can be passed
around as void*, do not have a DW_AT_type attribute. When building
type infos, just return "eris_typeinfo_pointer" if that is the case.