Fix accessing array values nested inside structs
The utility function cvalue_push() could not tell the difference between
having to push the array itself, or an item from the array. This patch
changes it so it only ever pushes to Lua a single value, and it does not
try to perform indexed array access; instead, the effective address
resulting of indexing an array is performed in eris_variable_index().
Coincidentally, this makes more sense than the previous behaviour, and
the code is easier to understand as well.