Fix handling of pointer values in cvalue_push()
Like with every other value, var->address contains the address where
the actual value is located, so in case of pointers, var->address is
the address of the pointer itself (e.g. for a T*, var->address then
a T**). This patch makes the code actually handle pointers that way.