Support copying ErisVariable values into the userdata
Sometimes it is desirable to be able to copy the value that will be referenced
by an ErisVariable userdata into the heap, as a payload of the ErisVariable
itself (like it is already done for data created from Lua by using __call on
an ErisTypeInfo). For example, after calling a C function which leaves its
return value into the stack, it is needed to copy it to the heap, otherwise
it will disappear under our feet once the current stack frame exits to return
the control to Lua.

This patch adds a boolean flag to eris_variable_push_userdata() which
instructs the function to allocate a payload with the size needed to hold a
value of the given type, and copy in there the value pointed by the passed
memory address. The effective address where the data is stored will be the
address of the ErisVariable payload.