fcall-ffi: Properly support pass-by-value semantics
This patch rewrites completely the way in which ErisTypeInfo values are mapped
to ffi_type* ones, in such a way that pass-by-value is supported for structs,
fixed-size arrays (where the size is known at compile-time), and combinations
of them embedded into each other.

To achieve this, we have to workaround the limitations of libffi in the
following ways:

1. Arrays cannot be represented as such using ffi_type, so instead we use a
   FFI_TYPE_STRUCT with as many elements as the array. This is a bit of a
   waste because of the repeated pointers ffi_type values for each item in
   the array, but it is tractable because usually we will be dealing with
   small-ish arrays.

2. Unions also cannot be represented using ffi_type, but as long as we
   instruct libffi of using the largest possible type that can be stored
   in the union, things just work, because fortunately libffi just does a
   memcpy() of the whole block instead of taking care to pick individual
   items. Lucky.

3. Arrays inside of structs (of sizes known at compile time) are embedded
   into the definition of the struct itself, because they contribute to the
   total size of the struct itself.

With those considerations, the current code is able to pass around arrays
inside structs which are part of an union, and other similary funny type
definitions, like the following real world example:

   struct Color {
     union {
       float rgba[4];
       struct {
         float r;
         float g;
         float b;
         float a;
       };
     };
   };

   struct Color RGBColor(uint8_t r, uint8_t g, uint8_t b) {
      return (struct Color) {
        .r = r / 255.f, .g = g / 255.f, .b = b / 255.f, .a = 1.0
      };
   }

Eris will happily call the RGBColor() function and receive a value of a
"struct Color", copy it into into the payload of an ErisVariable and push
the result to Lua.

Rock 'n' roll!  \m/ (>_<) \m/