Fix linking libffi with -Wl,--as-needed
The libraries to link must appear after -Wl,--as-needed in the order of linker
flags, otherwise libraries appearing before it will not be linked at build
time with a DT_NEEDED entry in the resulting ELF shared object. This patch
moves the libffi linker flags to the end of the flag list, fixing the issue.