Implement numeric indexing of function userdatas
Indexing a function userdata with a numeric index "n" retrieves the type
information for the "n"-th function parameter. This allows inspecting the
types of function parameters from Lua, which complements the existing ability
to retrieve the return type by indexing a function with "__type".

Additinoally, a test case is added which checks that the function return type,
and the types of its parameters can be obtained, and that invalid indexes
produce an error.