Make eol.sizeof() return the same values as LuaJIT's
The ffi.sizeof() function in LuaJIT returns "nil" for functions and "void",
and produces an error for any other value which is not either a Variable
userdata, or a CType userdata.

The behaviour of eol.sizeof() did not match that of ffi.sizeof(), even though
it was designed to behave in the same way. This patch corrects this defect.