Return "nil" from eol.type() when type is not found
Instead of raising an error when a type is not to be found in a library, return
"nil". This is done by checking whether accessing the DWARF debug information has
actually caused an error (in which case an error is still raised), or not (in
which case it means that looking up the TUE just failed because the type does
not exist).