Support dumping type cache statistics with TRACE()
This adds support for tracking misses and hits from the type cache. In order
to use it, configure Eol with --enable-trace, and the define the environment
variable EOL_TRACE=S. To filter only the cache stats, grep the string
"@type-cache" in the output of your program:

  EOL_TRACE=S lua program.lua 2>&1 | grep '@type-cache'

Output will be similat to the following:

  eol-module.c: @type-cache hits=0 misses=1
  eol-module.c: @type-cache hits=0 misses=2
  eol-module.c: @type-cache hits=0 misses=2
  eol-module.c: @type-cache hits=1 misses=9
  eol-module.c: @type-cache hits=2 misses=9
  eol-module.c: @type-cache hits=2 misses=13
  eol-module.c: @type-cache hits=144 misses=109
  eol-module.c: @type-cache hits=153 misses=109
  eol-module.c: @type-cache hits=155 misses=109
  eol-module.c: @type-cache hits=158 misses=109
  eol-module.c: @type-cache hits=160 misses=109
  eol-module.c: @type-cache hits=161 misses=109
  eol-module.c: @type-cache hits=166 misses=109
  eol-module.c: @type-cache hits=168 misses=109
  eol-module.c: @type-cache hits=169 misses=109
  eol-module.c: @type-cache hits=171 misses=109
  eol-module.c: @type-cache hits=173 misses=109

One trace line with statistics is printed right after each call from Lua to
the C functions which perform lookups in the cache. This is very useful to
evaluate the effectiveness of the cache by correlating the hit/miss
progression.